explicit uint32_t cast in pmurhash.c thanks to Radford Neal

+++ src/pmurhash.c
@@ -186,7 +186,7 @@ on big endian machines, or a byte-by-byte read if the endianess is unknown.
 #define DOBYTES(cnt, h1, c, n, ptr, len) do{ \
     int _i = cnt; \
     while(_i--) { \
-        c = c>>8 | *ptr++<<24; \
+        c = c>>8 | (uint32_t)*ptr++<<24; \
         n++; len--; \
         if(n==4) { \
             DOBLOCK(h1, c); \
